+++ src/main/java/com/hollingsworth/arsnouveau/common/entity/AbstractFlyingCreature.java
@@ -27,7 +27,7 @@ public abstract class AbstractFlyingCreature extends CreatureEntity {
         if (this.isInWater()) {
             this.moveRelative(0.02F, positionIn);
             this.move(MoverType.SELF, this.getDeltaMovement());
-            this.setDeltaMovement(this.getDeltaMovement().scale((double)0.8F));
+            this.setDeltaMovement(this.getDeltaMovement().scale(0.8F));
         } else if (this.isInLava()) {
             this.moveRelative(0.02F, positionIn);
             this.move(MoverType.SELF, this.getDeltaMovement());
@@ -47,7 +47,7 @@ public abstract class AbstractFlyingCreature extends CreatureEntity {
 
             this.moveRelative(this.onGround ? 0.1F * f1 : 0.02F, positionIn);
             this.move(MoverType.SELF, this.getDeltaMovement());
-            this.setDeltaMovement(this.getDeltaMovement().scale((double)f));
+            this.setDeltaMovement(this.getDeltaMovement().scale(f));
         }
 
         this.animationSpeedOld = this.animationSpeed;
