+++ src/main/java/com/hollingsworth/arsnouveau/common/entity/WildenStalker.java
@@ -178,7 +178,7 @@ public class WildenStalker extends MonsterEntity implements IAnimatable, IAnimat
         if (this.isInWater()) {
             this.moveRelative(0.02F, travelVector);
             this.move(MoverType.SELF, this.getDeltaMovement());
-            this.setDeltaMovement(this.getDeltaMovement().scale((double)0.8F));
+            this.setDeltaMovement(this.getDeltaMovement().scale(0.8F));
         } else if (this.isInLava()) {
             this.moveRelative(0.02F, travelVector);
             this.move(MoverType.SELF, this.getDeltaMovement());
@@ -198,7 +198,7 @@ public class WildenStalker extends MonsterEntity implements IAnimatable, IAnimat
 
             this.moveRelative(this.onGround ? 0.1F * f1 : 0.02F, travelVector);
             this.move(MoverType.SELF, this.getDeltaMovement());
-            this.setDeltaMovement(this.getDeltaMovement().scale((double)f));
+            this.setDeltaMovement(this.getDeltaMovement().scale(f));
         }
 
         this.calculateEntityAnimation(this, false);
