+++ src/main/java/com/hollingsworth/arsnouveau/common/potions/SnareEffect.java
@@ -11,7 +11,7 @@ public class SnareEffect extends Effect {
     public SnareEffect() {
         super(EffectType.HARMFUL, 2039587);
         setRegistryName(ArsNouveau.MODID, "snared");
-        addAttributeModifier(Attributes.MOVEMENT_SPEED, "0dee8a21-f182-42c8-8361-1ad6186cac30", (double)-1, AttributeModifier.Operation.MULTIPLY_TOTAL);
+        addAttributeModifier(Attributes.MOVEMENT_SPEED, "0dee8a21-f182-42c8-8361-1ad6186cac30", -1, AttributeModifier.Operation.MULTIPLY_TOTAL);
     }
 
 }
